Extract MemberPage to its own file
This step is in preparation of testing MemberPage data with scraper_test
